requirements.txt
asgiref==3.7.2
# aaertifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
defusedxml==0.7.1
Django==5.0.1
django-admin-thumbnails==0.2.8
django-allauth==0.60.1
django-ckeditor==6.7.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-taggit==5.0.1
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.3
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wheel==0.42.0
